station/requirements.txt
wrapt == 1.14.1
PyYAML == 6.0
paho-mqtt == 1.6.1